backend/requirements.txt
Django>=4.1.2
psycopg2>=2.9.4
sib-api-v3-sdk>=7.5.0
django-environ>=0.9.0
djangorestframework>=3.14.0
django-cors-headers>=3.13.0
ortools>=9.4.1874
sentry-sdk>=1.10.1
